--- components/ArtworkPreview/index.js
@@ -1,3 +1,0 @@
-export default function ArtworkPreview({ item }) {
-  return;
-}
