+++ components/Artworks/index.js
@@ -1,38 +1,6 @@
 import Image from "next/image";
-import useSWR from "swr";
-import ArtworkPreview from "../ArtworkPreview/index.js";
-import ArtworkInfo from "../ArtworkInfo/index.js";
 
-const fetcher = async (url) => {
-  const res = await fetch(url);
-
-  if (!res.ok) {
-    const error = new Error("An error occurred while fetching the data.");
-    error.info = await res.json();
-    error.status = res.status;
-    throw error;
-  }
-
-  return res.json();
-};
-
-export default function Artworks() {
-  const {
-    data: artworks,
-    error,
-    isLoading,
-  } = useSWR(
-    "https://api.smk.dk/api/v1/art/search/?keys=*&fields=image_thumbnail&fields=titles&fields=id&fields=production&fields=dimensions&fields=current_location_name&fields=production_dates_notes&fields=labels&filters=[image_hq:true],[object_names:painting],[public_domain:true]&offset=0&rows=2000&lang=en",
-    fetcher
-  );
-
-  if (error) {
-    return <div>failed to load</div>;
-  }
-  if (isLoading) {
-    return <div>loading...</div>;
-  }
-
+export default function Artworks({ artworks }) {
   return (
     console.log(artworks),
     (
